+++ src/components/targets-map/TargetsMap.js
@@ -1,8 +1,6 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Map, TileLayer, Marker, CircleMarker } from 'react-leaflet';
-import { useSelector } from 'react-redux';
 import { func, string } from 'prop-types';
-import { useDispatch } from 'react-redux';
 
 import 'components/targets-map/targets-map.scss';
 import {
@@ -14,25 +12,17 @@ import { mapConstants } from 'constants/map.constants';
 import { newTargetIcon } from 'components/targets-map/map-icons';
 import MarketTarget from 'components/targets-map/MarkerTargets';
 import { latLngShape } from 'constants/shapes';
-import { userActions } from 'actions/user.actions';
+import useTargetsMap from 'hooks/useTargetsMap';
 
 const TargetsMap = ({
   newTargetlatlng,
   setNewTargetlatlng,
   newTargetRadius,
 }) => {
-  const { getTargets } = userActions;
-
-  const dispatch = useDispatch();
-
   const createTarget = (e) => {
     setNewTargetlatlng(e.latlng);
   };
-  const { userTargets } = useSelector((state) => state.target);
-
-  useEffect(() => {
-    dispatch(getTargets());
-  }, [dispatch, getTargets]);
+  const { userTargets } = useTargetsMap();
 
   return (
     <Map
